requirements.txt
discord.py==2.2.3
requests
googlesearch-python
beautifulsoup4